Added examples of ResponseEntity, Basic Thymeleaf and Spring API client

+++ src/main/java/com/alreylz/springbootcourse/controllers/ComputerRestControllerWResponseEntity.java
@@ -3,19 +3,23 @@ package com.alreylz.springbootcourse.controllers;
 import com.alreylz.springbootcourse.mysql.entities.Computer;
 import com.alreylz.springbootcourse.repos.ComputerDAO;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+
 /**
- * RESTFUL API para operaciones CRUD sobre la entidad Computer === tabla 'computers'
- * Utiliza un DAO que se crea a partir de CrudRepository<Computer,Long>
+ * Clase Controller que además de exponer endpoints para crear, leer, actualizar y eliminar registros 'Computer',
+ * hace uso del contenedor ResponseEntity para retornar códigos de error informativos al cliente del microservicio
+ * (es una copia 'mejorada' de ComputerRestController
  */
 @Controller
-public class ComputerRestController {
+public class ComputerRestControllerWResponseEntity {
 
-    static final String basePath = "/restful";
+    static final String basePath = "/restfulRespEntity";
 
     // DAO (Data Access Object) / CRUD
     @Autowired
@@ -23,34 +27,41 @@ public class ComputerRestController {
 
     // GET - Extraer info de bbdd (listar todos los ordenadores)
     @RequestMapping(value = basePath + "/computers", method = RequestMethod.GET)
-    public @ResponseBody
-    List<Computer> getAllComputers() {
+    ResponseEntity<List<Computer>> getAllComputers() {
         List<Computer> computerList = null;
+
         try {
             computerList = cDao.findAll();
 
+            if (computerList.isEmpty() || computerList == null) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } catch (Exception e) {
             System.out.println(e);
-
-
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-        return computerList;
+
+
+        return new ResponseEntity<>(computerList, HttpStatus.OK);
     }
 
 
     // GET (by id) - Extraer info de bbdd (listar un ordenador por id)
     @RequestMapping(value = basePath + "/computers/{id}", method = RequestMethod.GET)
     public @ResponseBody
-    Computer getComputer(@PathVariable String id) {
+    ResponseEntity<Computer> getComputer(@PathVariable String id) {
         Computer fromDbComputer = null;
         try {
             Long longId = Long.parseLong(id);
             fromDbComputer = cDao.findById(longId).orElse(null);
-
+            if (fromDbComputer == null) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } catch (Exception e) {
             System.out.println(e);
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-        return fromDbComputer;
+        return new ResponseEntity<>(fromDbComputer, HttpStatus.OK);
     }
 
 
@@ -59,19 +70,27 @@ public class ComputerRestController {
     // -> @RequestParam por contra, requeriría extraer uno a uno los parámetros del JSON que viene
     @RequestMapping(value = basePath + "/computers", method = RequestMethod.POST)
     public @ResponseBody
-    Computer insertComputer(@RequestBody Computer c) {
+    ResponseEntity<Computer> insertComputer(@RequestBody Computer c) {
+
+        if (c != null) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         try {
             cDao.save(c);
         } catch (Exception e) {
             System.out.println(e);
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+
         }
-        return c;
+        return new ResponseEntity<>(c, HttpStatus.OK);
+
     }
 
     // PUT - Actualización de info en BBDD
     @RequestMapping(value = basePath + "/computers/{id}", method = RequestMethod.PUT)
     public @ResponseBody
-    Computer updateComputer(@PathVariable String id, @RequestBody Computer toUpdate) {
+    ResponseEntity<Computer> updateComputer(@PathVariable String id, @RequestBody Computer toUpdate) {
         Computer updatedComputer = null;
         try {
             //Convertimos el id al tipo de datos del id de base de datos (Long integer)
@@ -82,29 +101,41 @@ public class ComputerRestController {
             if (fromDBComputer != null) {
                 cDao.save(toUpdate);
                 updatedComputer = toUpdate;
+
+            } else {
+                return new ResponseEntity<>(updatedComputer, HttpStatus.BAD_REQUEST);
             }
         } catch (Exception e) {
             System.out.println(e);
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+
         }
-        return updatedComputer;
+        return new ResponseEntity<>(updatedComputer, HttpStatus.OK);
     }
 
     // DELETE - Eliminación de info en BBDD
     @RequestMapping(value = basePath + "/computers/{id}", method = RequestMethod.DELETE)
     public @ResponseBody
-    Computer deleteOne(@PathVariable String id) {
+    ResponseEntity<Computer> deleteOne(@PathVariable String id) {
         Computer toDeleteComputer = null;
         try {
             //Convertimos el id al tipo de datos del id de base de datos (Long integer)
             Long intId = Long.parseLong(id);
             // Buscamos por id y en caso de no encontrar nada, obtenemos null
             toDeleteComputer = cDao.findById(intId).orElse(null);
+            if (toDeleteComputer == null)
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+
+
             cDao.delete(toDeleteComputer);
         } catch (Exception e) {
             System.out.println(e);
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
-        return toDeleteComputer;
+
+        return new ResponseEntity<>(toDeleteComputer, HttpStatus.OK);
     }
 
 
 }
+
